dev-data-bridge-agent: Renamed classes and reorganised subpackages
